split Datepicker into index and render
