+++ src/lib/Datepicker/index.js
@@ -1,11 +1,12 @@
-import moment from './mixins/moment'
+import moment from '../mixins/moment'
 import { flatten } from 'lodash'
-import { getCalendarMatrix, isBetween } from './helpers/dates'
-import { warn } from '../util'
+import { getCalendarMatrix, isBetween } from '../helpers/dates'
+import render from './render'
 
 export default {
   name: 'VkDatepicker',
   mixins: [moment],
+  render,
   props: {
     dates: {
       type: Array,
@@ -24,18 +25,6 @@ export default {
     max: {
       type: [String, Number],
       default: '2050-12-31'
-    }
-  },
-  render (h) {
-    if (this.$slots.default !== undefined) {
-      const calendarVnode = this.$slots.default[0]
-      calendarVnode.componentOptions.propsData.fieldComponent = DateField
-      calendarVnode.componentOptions.propsData.fieldProps = {
-        $datepicker: this
-      }
-      return calendarVnode
-    } else {
-      warn('VkDatepicker expects VkCalendar as child.')
     }
   },
   computed: {
@@ -77,35 +66,5 @@ export default {
         ? this.$emit('unpick', moment)
         : this.$emit('pick', moment)
     }
-  }
-}
-
-const DateField = {
-  name: 'DateField',
-  functional: true,
-  props: {
-    date: {
-      required: true
-    },
-    $datepicker: {
-      required: true
-    }
-  },
-  render (h, { parent: $calendar, props }) {
-    const $datepicker = props.$datepicker
-    return (
-      <a class={{
-        'uk-active': $datepicker.isPicked(props.date),
-        'uk-datepicker-table-disabled': $datepicker.isDisabled(props.date),
-        'uk-datepicker-table-muted': !$calendar.isInCurrentMonth(props.date) ||
-          $datepicker.isDisabled(props.date)
-      }} on-click={() => {
-        if (!$datepicker.isDisabled(props.date)) {
-          $datepicker.toggle(props.date)
-        }
-      }}>
-        { props.date.format('D') }
-      </a>
-    )
   }
 }
